[lib/c] fixes for changed implementation of fibonacci

+++ src/002-even-fibonacci-numbers/c/src/solve.h
@@ -1,5 +1,4 @@
 #include <stdint.h>
-#include <euler/math.h>
 #pragma once
 
-int solve(int max);
+uint64_t solve(uint64_t max);
